refactor: 클래스 의존성 방향 재설정 및 주차장 생성 이벤트 전달값 수정
