+++ src/main/java/com/parkmate/parkingreadservice/parkinglotread/domain/ParkingLotOption.java
@@ -1,0 +1,14 @@
+package com.parkmate.parkingreadservice.parkinglotread.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ParkingLotOption {
+
+    private String name;
+    private String label;
+}
