+++ src/main/java/com/parkmate/parkingreadservice/parkinglotread/infrastructure/CustomMongoRepository.java
@@ -1,7 +1,7 @@
 package com.parkmate.parkingreadservice.parkinglotread.infrastructure;
 
-import com.parkmate.parkingreadservice.parkinglotread.event.ParkingLotMetadataUpdateEvent;
-import com.parkmate.parkingreadservice.parkinglotread.event.ParkingLotReactionsUpdateEvent;
+import com.parkmate.parkingreadservice.kafka.event.ParkingLotMetadataUpdateEvent;
+import com.parkmate.parkingreadservice.kafka.event.ParkingLotReactionsUpdateEvent;
 
 import java.util.List;
