+++ src/main/java/com/parkmate/parkingreadservice/parkinglotread/infrastructure/CustomMongoRepositoryImpl.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mongodb.MongoBulkWriteException;
 import com.parkmate.parkingreadservice.parkinglotread.domain.ParkingLotRead;
-import com.parkmate.parkingreadservice.parkinglotread.event.ParkingLotMetadataUpdateEvent;
-import com.parkmate.parkingreadservice.parkinglotread.event.ParkingLotReactionsUpdateEvent;
-import com.parkmate.parkingreadservice.parkinglotread.event.ReactionType;
+import com.parkmate.parkingreadservice.kafka.event.ParkingLotMetadataUpdateEvent;
+import com.parkmate.parkingreadservice.kafka.event.ParkingLotReactionsUpdateEvent;
+import com.parkmate.parkingreadservice.kafka.event.ReactionType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.mongodb.core.BulkOperations;
