+++ src/main/java/com/parkmate/parkingreadservice/parkinglotread/presentation/ParkingLotReadController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/parking-lot-read")
+@RequestMapping("/api/v1/parkingLots")
 @RequiredArgsConstructor
 public class ParkingLotReadController {
